@@ -51,9 +51,9 @@
     "jquery.scrollto": "^2.1.2",
     "prop-types": "^15.6.0",
     "querystring": "^0.2.0",
-    "react": "^16.3.2",
+    "react": "^16.8.0",
     "react-class": "^3.2.2",
-    "react-dom": "^16.3.2",
+    "react-dom": "^16.8.0",
     "react-jsonschema-form": "^1.0.0",
     "redis": "^2.6.0-2",
     "request": "^2.72.0",
