@@ -11,7 +11,7 @@ logger = logging.getLogger(__name__)
 
 if __name__ == '__main__':
     if not MULTISERVER_ENABLED:
-        logger.error("MULTISERVER_ENABLED is not set.  Exiting")
+        logger.error(u"MULTISERVER_ENABLED is not set.  Exiting")
         exit()
     sched = run_background_scheduler()
     sched.print_jobs()
