@@ -36,13 +36,13 @@ BS_CAPS = [
 
 SAUCE_CORE_CAPS = [
     {'browserName': "firefox", 'platform': "macOS 10.12", 'version': "latest", "screenResolution": "1920x1440",
-     'sefaria_mode': 'multi_panel', 'sefaria_short_name': 'FF/x12'},
+     'sefaria_mode': 'multi_panel', 'sefaria_short_name': 'FF/x12', "extendedDebugging": True},
 
     {'browserName': "safari", 'platform': "OS X 10.11", 'version': "latest", "screenResolution": "1920x1440",
      'sefaria_mode': 'multi_panel', 'sefaria_short_name': 'Sf/x11'},
 
     {'browserName': "chrome", 'platform': "Windows 10", 'version': "latest", "screenResolution": "1920x1080",
-     'sefaria_mode': 'multi_panel', 'sefaria_short_name': 'Cr/w10'},
+     'sefaria_mode': 'multi_panel', 'sefaria_short_name': 'Cr/w10', "extendedDebugging": True},
 
     #{'browserName': "firefox", 'platform': "Windows 10", 'version': "latest", "screenResolution": "1920x1080",
     #    'sefaria_mode': 'multi_panel', 'sefaria_short_name': 'FF/w10'},
