@@ -38,7 +38,7 @@ from django.utils import timezone
 from sefaria.model import *
 from sefaria.workflows import *
 from sefaria.reviews import *
-from sefaria.model.user_profile import user_link, user_started_text, unread_notifications_count_for_user, public_user_data
+from sefaria.model.user_profile import user_link, user_started_text, unread_notifications_count_for_user, public_user_data, user_stats_data, site_stats_data
 from sefaria.model.group import GroupSet
 from sefaria.model.topic import get_topics
 from sefaria.model.webpage import get_webpages_for_ref
@@ -50,7 +50,7 @@ from sefaria.client.util import jsonResponse
 from sefaria.history import text_history, get_maximal_collapsed_activity, top_contributors, make_leaderboard, make_leaderboard_condition, text_at_revision, record_version_deletion, record_index_deletion
 from sefaria.system.decorators import catch_error_as_json, sanitize_get_params, json_response_decorator
 from sefaria.summaries import get_or_make_summary_node
-from sefaria.sheets import get_sheets_for_ref, public_sheets, get_sheets_by_tag, user_sheets, user_tags, recent_public_tags, sheet_to_dict, get_top_sheets, public_tag_list, group_sheets, get_sheet_for_panel, annotate_user_links
+from sefaria.sheets import get_sheets_for_ref, public_sheets, get_sheets_by_tag, user_sheets, user_tags, trending_tags, sheet_to_dict, get_top_sheets, public_tag_list, group_sheets, get_sheet_for_panel, annotate_user_links
 from sefaria.utils.util import list_depth, text_preview
 from sefaria.utils.hebrew import hebrew_plural, hebrew_term, encode_hebrew_numeral, encode_hebrew_daf, is_hebrew, strip_cantillation, has_cantillation
 from sefaria.utils.talmud import section_to_daf, daf_to_section
@@ -73,7 +73,7 @@ logger = logging.getLogger(__name__)
 
 #    #    #
 # Initialized cache library objects that depend on sefaria.model being completely loaded.
-logger.warn("Initializing library objects.")
+logger.warn(u"Initializing library objects.")
 library.get_toc_tree()
 library.build_full_auto_completer()
 library.build_ref_auto_completer()
@@ -632,7 +632,7 @@ def sheets(request):
         "initialMenu": "sheets",
         "topSheets": get_top_sheets(),
         "tagList": public_tag_list(sort_by="count"),
-        "trendingTags": recent_public_tags(days=14, ntags=18)
+        "trendingTags": trending_tags(ntags=18)
     })
 
     title = _("Sefaria Source Sheets")
@@ -814,7 +814,7 @@ def topics_page(request):
         "initialMenu":  "topics",
         "initialTopic": None,
         "topicList": topics.list(sort_by="count"),
-        "trendingTags": recent_public_tags(days=14, ntags=12),
+        "trendingTags": trending_tags(ntags=12),
     })
 
     propsJSON = json.dumps(props)
@@ -907,12 +907,25 @@ def updates(request):
     desc  = _("See texts, translations and connections that have been recently added to Sefaria.")
     return menu_page(request, props, "updates", title, desc)
 
-@login_required
+
+def new_home(request):
+    props = base_props(request)
+    title = _("Sefaria Stories")
+    return menu_page(request, props, "homefeed", title)
+
+
+@staff_member_required
 def story_editor(request):
     props = base_props(request)
     title = _("Story Editor")
     return menu_page(request, props, "story_editor", title)
 
+
+@staff_member_required
+def user_stats(request):
+    props = base_props(request)
+    title = _("User Stats")
+    return menu_page(request, props, "user_stats", title)
 
 @login_required
 def account(request):
@@ -1100,8 +1113,9 @@ def edit_text_info(request, title=None, new_title=None):
         # Add New
         new_title = new_title.replace("_", " ")
         try: # Redirect to edit path if this title already exists
+            library.get_index(new_title)
             return redirect("/edit/textinfo/%s" % new_title)
-        except:
+        except BookNameError:
             pass
         indexJSON = json.dumps({"title": new_title})
         text_exists = False
@@ -2469,8 +2483,8 @@ def stories_api(request, gid=None):
 
             payload = json.loads(request.POST.get("json"))
             try:
-                SharedStory(payload).save()
-                return jsonResponse({"status": "ok"})
+                s = SharedStory(payload).save()
+                return jsonResponse({"status": "ok", "story": s.contents()})
             except AssertionError as e:
                 return jsonResponse({"error": e.message})
 
@@ -2479,8 +2493,8 @@ def stories_api(request, gid=None):
             def protected_post(request):
                 payload = json.loads(request.POST.get("json"))
                 try:
-                    SharedStory(payload).save()
-                    return jsonResponse({"status": "ok"})
+                    s = SharedStory(payload).save()
+                    return jsonResponse({"status": "ok", "story": s.contents()})
                 except AssertionError as e:
                     return jsonResponse({"error": e.message})
 
@@ -2538,6 +2552,22 @@ def addDynamicStories(stories, user, page):
             stories = [stry.contents()] + stories
 
     return stories
+
+
+@staff_member_required
+def user_stats_api(request, uid):
+    assert request.method == "GET", "Unsupported Method"
+    quick = bool(request.GET.get("quick", False))
+    if quick:
+        return jsonResponse(public_user_data(uid))
+    # Todo: or now, we're hard-coding Rosh Hashannah 2018.
+    return jsonResponse(user_stats_data(uid, start=datetime(2018, 9, 9)))
+
+
+@staff_member_required
+def site_stats_api(request):
+    assert request.method == "GET", "Unsupported Method"
+    return jsonResponse(site_stats_data())
 
 
 @staff_member_required
@@ -3049,60 +3079,38 @@ def leaderboard(request):
 
 @ensure_csrf_cookie
 @sanitize_get_params
-def user_profile(request, username, page=1):
+def user_profile(request, username):
     """
     User's profile page.
     """
     try:
-        profile    = UserProfile(slug=username)
+        profile = UserProfile(slug=username)
     except Exception, e:
         # Couldn't find by slug, try looking up by username (old style urls)
         # If found, redirect to new URL
         # If we no longer want to support the old URLs, we can remove this
-        user       = get_object_or_404(User, username=username)
-        profile    = UserProfile(id=user.id)
+        user = get_object_or_404(User, username=username)
+        profile = UserProfile(id=user.id)
 
         return redirect("/profile/%s" % profile.slug, permanent=True)
 
+    props = base_props(request)
+    profileJSON = profile.to_api_dict()
+    props.update({
+        "initialMenu":  "profile",
+        "initialProfile": profileJSON,
+    })
+    title = u"%(full_name)s on Sefaria" % {"full_name": profile.full_name}
+    desc = u'%(full_name)s is on Sefaria. Follow to view their public source sheets, notes and translations.' % {"full_name": profile.full_name}
 
-    following      = profile.followed_by(request.user.id) if request.user.is_authenticated else False
-
-    page_size      = 20
-    page           = int(page) if page else 1
-    if page > 40:
-        generic_response = { "title": "Activity Unavailable", "content": "You have requested a page deep in Sefaria's history.<br><br>For performance reasons, this page is unavailable. If you need access to this information, please <a href='mailto:dev@sefaria.org'>email us</a>." }
-        return render(request,'static/generic.html', generic_response)
-
-    query          = {"user": profile.id}
-    filter_type    = request.GET["type"] if "type" in request.GET else None
-    activity, apage= get_maximal_collapsed_activity(query=query, page_size=page_size, page=page, filter_type=filter_type)
-    notes, npage   = get_maximal_collapsed_activity(query=query, page_size=page_size, page=page, filter_type="add_note")
-
-    contributed    = activity[0]["date"] if activity else None
-    scores         = db.leaders_alltime.find_one({"_id": profile.id})
-    score          = int(scores["count"]) if scores else 0
-    user_texts     = scores.get("texts", None) if scores else None
-    sheets         = db.sheets.find({"owner": profile.id, "status": "public"}, {"id": 1, "datePublished": 1}).sort([["datePublished", -1]])
-
-    next_page      = apage + 1 if apage else None
-    next_page      = "/profile/%s/%d" % (username, next_page) if next_page else None
-
-    return render(request,"profile.html",
-                             {
-                                'profile': profile,
-                                'following': following,
-                                'activity': activity,
-                                'sheets': sheets,
-                                'notes': notes,
-                                'joined': profile.date_joined,
-                                'contributed': contributed,
-                                'score': score,
-                                'scores': scores,
-                                'user_texts': user_texts,
-                                'filter_type': filter_type,
-                                'next_page': next_page,
-                                "single": False,
-                              })
+    propsJSON = json.dumps(props)
+    html = render_react_component("ReaderApp", propsJSON)
+    return render(request,'base.html', {
+        "propsJSON":      propsJSON,
+        "title":          title,
+        "desc":           desc,
+        "html":           html,
+    })
 
 
 @catch_error_as_json
@@ -3128,7 +3136,25 @@ def profile_api(request):
             return jsonResponse({"error": error})
         else:
             profile.save()
-            return jsonResponse(profile.to_DICT())
+            return jsonResponse(profile.to_mongo_dict())
+    return jsonResponse({"error": "Unsupported HTTP method."})
+
+
+@catch_error_as_json
+def profile_get_api(request, slug):
+    if request.method == "GET":
+        profile = UserProfile(slug=slug)
+        return jsonResponse(profile.to_api_dict())
+    return jsonResponse({"error": "Unsupported HTTP method."})
+
+
+@catch_error_as_json
+def profile_follow_api(request, ftype, slug):
+    if request.method == "GET":
+        profile = UserProfile(slug=slug)
+        follow_set = FollowersSet(profile.id) if ftype == "followers" else FolloweesSet(profile.id)
+        response = [UserProfile(id=uid).to_api_dict(basic=True) for uid in follow_set.uids]
+        return jsonResponse(response)
     return jsonResponse({"error": "Unsupported HTTP method."})
 
 
@@ -3288,11 +3314,13 @@ def account_settings(request):
                                 'profile': profile,
                               })
 
+
 @login_required
 def enable_home_feed(request):
     resp = home(request, True)
     resp.set_cookie("home_feed", "yup", 60 * 60 * 24 * 365)
     return resp
+
 
 @login_required
 def disable_home_feed(request):
@@ -3310,19 +3338,7 @@ def home(request, show_feed=None):
         show_feed = request.COOKIES.get("home_feed", None)
 
     if show_feed:
-        props = base_props(request)
-
-        props.update({
-            "initialMenu": "homefeed"
-        })
-        propsJSON = json.dumps(props)
-        html = render_react_component("ReaderApp", propsJSON)
-        return render(request, 'base.html', {
-            "propsJSON": propsJSON,
-            "html": html,
-            "title": "Sefaria Stories",
-            "desc": "",
-        })
+        return redirect("/new-home")
 
     if not SITE_SETTINGS["TORAH_SPECIFIC"]:
         return redirect("/texts")
@@ -3804,17 +3820,13 @@ def random_by_topic_api(request):
     Returns Texts API data for a random text taken from popular topic tags
     """
     cb = request.GET.get("callback", None)
-    topics_filtered = filter(lambda x: x['count'] > 15, get_topics().list())
+    topics_filtered = filter(lambda x: x['good_to_promote'], get_topics().list())
     if len(topics_filtered) == 0:
         resp = jsonResponse({"ref": None, "topic": None, "url": None}, callback=cb)
         resp['Content-Type'] = "application/json; charset=utf-8"
         return resp
     random_topic = choice(topics_filtered)['tag']
     term = Term().load_by_title(random_topic)
-    if term is not None and getattr(term, "sensitive", False):
-        # term is sensitive, try again
-        return random_by_topic_api(request)
-        
     random_source = choice(get_topics().get(random_topic).contents()['sources'])[0]
     try:
         oref = Ref(random_source)
