@@ -2,7 +2,7 @@
 dependencies.py -- list cross model dependencies and subscribe listeners to changes.
 """
 
-from . import abstract, link, note, history, schema, text, layer, version_state, translation_request, timeperiod, person, garden, notification, story, group, library, category
+from . import abstract, link, note, history, schema, text, layer, version_state, translation_request, timeperiod, person, garden, notification, story, group, library, category, ref_data, user_profile
 
 from abstract import subscribe, cascade, cascade_to_list, cascade_delete, cascade_delete_to_list
 import sefaria.system.cache as scache
@@ -15,13 +15,17 @@ subscribe(text.process_index_change_in_toc,                             text.Ind
 
 # Index Name Change
 subscribe(text.process_index_title_change_in_core_cache,                text.Index, "attributeChange", "title")
+subscribe(text.process_index_title_change_in_versions,                  text.Index, "attributeChange", "title")
+subscribe(version_state.process_index_title_change_in_version_state,    text.Index, "attributeChange", "title")
 subscribe(link.process_index_title_change_in_links,                     text.Index, "attributeChange", "title")
 subscribe(note.process_index_title_change_in_notes,                     text.Index, "attributeChange", "title")
 subscribe(history.process_index_title_change_in_history,                text.Index, "attributeChange", "title")
-subscribe(text.process_index_title_change_in_versions,                  text.Index, "attributeChange", "title")
 subscribe(text.process_index_title_change_in_dependant_records,         text.Index, "attributeChange", "title")
-subscribe(version_state.process_index_title_change_in_version_state,    text.Index, "attributeChange", "title")
+subscribe(text.process_index_title_change_in_sheets,                    text.Index, "attributeChange", "title")
 subscribe(cascade(notification.GlobalNotificationSet, "content.index"), text.Index, "attributeChange", "title")
+subscribe(ref_data.process_index_title_change_in_ref_data,              text.Index, "attributeChange", "title")
+subscribe(user_profile.process_index_title_change_in_user_history,      text.Index, "attributeChange", "title")
+
 # Taken care of on save
 # subscribe(text.process_index_change_in_toc,                             text.Index, "attributeChange", "title")
 
@@ -35,6 +39,7 @@ subscribe(text.process_index_delete_in_versions,                        text.Ind
 subscribe(translation_request.process_index_delete_in_translation_requests, text.Index, "delete")
 subscribe(text.process_index_delete_in_toc,                             text.Index, "delete")
 subscribe(cascade_delete(notification.GlobalNotificationSet, "content.index", "title"),   text.Index, "delete")
+subscribe(ref_data.process_index_delete_in_ref_data,                    text.Index, "delete")
 
 
 # Process in ES
