@@ -299,7 +299,7 @@ class Garden(abst.AbstractMongoRecord):
 
         sheet = Sheet().load({"id": sheet_id})
         if not sheet:
-            logger.warning("Failed to load sheet {}".format(sheet_id))
+            logger.warning(u"Failed to load sheet {}".format(sheet_id))
 
         def process_sources(sources, tags):
             for source in sources:
@@ -543,7 +543,7 @@ def process_index_title_change_in_gardens(indx, **kwargs):
         try:
             l.save()
         except InputError: #todo: this belongs in a better place - perhaps in abstract
-            logger.warning("Deleting link that failed to save: {} {}".format(l.refs[0], l.refs[1]))
+            logger.warning(u"Deleting link that failed to save: {} {}".format(l.refs[0], l.refs[1]))
             l.delete()
 
 
