@@ -11,7 +11,6 @@ from sefaria.system.database import db
 from . import abstract as abst
 from . import text
 
-
 import logging
 logger = logging.getLogger(__name__)
 
@@ -29,8 +28,8 @@ class Link(abst.AbstractMongoRecord):
     ]
 
     optional_attrs = [
-        "expandedRefs0",  # list of refs corresponding to `refs.0`, but breaking ranging refs down into individual segments
-        "expandedRefs1",  # list of refs corresponding to `refs.1`, but breaking ranging refs down into individual segments
+        "expandedRefs0",    # list of refs corresponding to `refs.0`, but breaking ranging refs down into individual segments
+        "expandedRefs1",    # list of refs corresponding to `refs.1`, but breaking ranging refs down into individual segments
         "anchorText",       # string of dibbur hamatchil (largely depcrated) 
         "availableLangs",   # list of lists corresponding to `refs` showing languages available, e.g. [["he"], ["he", "en"]]  
         "auto",             # bool whether generated by automatic process
@@ -153,8 +152,9 @@ class LinkSet(abst.AbstractMongoSet):
         '''
         try:
             regex_list = query_or_ref.regex(as_list=True)
-            ref_clauses = [{"refs": {"$regex": r}} for r in regex_list]
-            super(LinkSet, self).__init__({"$or": ref_clauses}, page, limit, hint=[("refs", 1)])
+            ref_clauses = [{"expandedRefs0": {"$regex": r}} for r in regex_list]
+            ref_clauses += [{"expandedRefs1": {"$regex": r}} for r in regex_list]
+            super(LinkSet, self).__init__({"$or": ref_clauses}, page, limit)
         except AttributeError:
             super(LinkSet, self).__init__(query_or_ref, page, limit)
 
@@ -274,10 +274,14 @@ def process_index_title_change_in_links(indx, **kwargs):
     links = LinkSet({"$or": queries})
     for l in links:
         l.refs = [r.replace(kwargs["old"], kwargs["new"], 1) if re.search(u'|'.join(patterns), r) else r for r in l.refs]
+        l.expandedRefs0 = [r.replace(kwargs["old"], kwargs["new"], 1) if re.search(u'|'.join(patterns), r) else r for r in l.expandedRefs0]
+        l.expandedRefs1 = [r.replace(kwargs["old"], kwargs["new"], 1) if re.search(u'|'.join(patterns), r) else r for r in l.expandedRefs1]
         try:
+            l._skip_lang_check = True
+            l._skip_expanded_refs_set = True
             l.save()
         except InputError: #todo: this belongs in a better place - perhaps in abstract
-            logger.warning("Deleting link that failed to save: {} - {}".format(l.refs[0], l.refs[1]))
+            logger.warning(u"Deleting link that failed to save: {} - {}".format(l.refs[0], l.refs[1]))
             l.delete()
 
 
@@ -297,6 +301,9 @@ def update_link_language_availabiliy(oref, lang=None, available=None):
     
     if lang and available is not None:
         for link in links:
+            if not getattr(link, "availableLangs", None):
+                link.save()
+                continue
             pos = 0 if oref.overlaps(text.Ref(link.refs[0])) else 1
 
             if available:
