@@ -4,10 +4,12 @@ Writes to MongoDB Collection: notes
 """
 
 import regex as re
-import bleach
 
 from . import abstract as abst
-from sefaria.model.text import Ref, IndexSet
+from sefaria.model.text import Ref
+
+import logging
+logger = logging.getLogger(__name__)
 
 
 class Note(abst.AbstractMongoRecord):
@@ -53,7 +55,7 @@ def process_index_title_change_in_notes(indx, **kwargs):
             n.ref = n.ref.replace(kwargs["old"], kwargs["new"], 1)
             n.save()
         except Exception:
-            logger.warning("Deleting note that failed to save: {}".format(n.ref))
+            logger.warning(u"Deleting note that failed to save: {}".format(n.ref))
             n.delete()
 
 def process_index_delete_in_notes(indx, **kwargs):
