@@ -1,6 +1,11 @@
+import math
+
 from . import abstract as abst
 from . import text
-import math
+from sefaria.system.exceptions import InputError
+
+import logging
+logger = logging.getLogger(__name__)
 
 
 class RefData(abst.AbstractMongoRecord):
@@ -38,3 +43,26 @@ class RefDataSet(abst.AbstractMongoSet):
 
     def top_ref(self):
         return self.nth_ref(1)
+
+
+def process_index_title_change_in_ref_data(indx, **kwargs):
+    print "Cascading Ref Data from {} to {}".format(kwargs['old'], kwargs['new'])
+
+    # ensure that the regex library we're using here is the same regex library being used in `Ref.regex`
+    from text import re as reg_reg
+    patterns = [pattern.replace(reg_reg.escape(indx.title), reg_reg.escape(kwargs["old"]))
+                for pattern in text.Ref(indx.title).regex(as_list=True)]
+    queries = [{'ref': {'$regex': pattern}} for pattern in patterns]
+    objs = RefDataSet({"$or": queries})
+    for o in objs:
+        o.ref = o.ref.replace(kwargs["old"], kwargs["new"], 1)
+        try:
+            o.save()
+        except InputError:
+            logger.warning(u"Failed to convert ref data from: {} to {}".format(kwargs['old'], kwargs['new']))
+
+
+def process_index_delete_in_ref_data(indx, **kwargs):
+    from sefaria.model.text import prepare_index_regex_for_dependency_process
+    pattern = prepare_index_regex_for_dependency_process(indx)
+    RefDataSet({"ref": {"$regex": pattern}}).delete()
