@@ -487,7 +487,7 @@ class TextPassageStoryFactory(AbstractStoryFactory):
             "ref": oref.normal(),
             "title": kwargs.get("title", {"en": oref.normal(), "he": oref.he_normal()})
         }
-        if kwargs.get("lead"):
+        if kwargs.get("lead") and kwargs.get("lead").get("en") and kwargs.get("lead").get("he"):
             d["lead"] = kwargs.get("lead")
 
         if kwargs.get("versions"):
@@ -824,7 +824,7 @@ class GroupSheetListFactory(AbstractStoryFactory):
             "title": kwargs.get("title",{"en": g.name, "he": g.name}),
             "cozy": kwargs.get("cozy", False)
         }
-        if kwargs.get("lead"):
+        if kwargs.get("lead") and kwargs.get("lead").get("en") and kwargs.get("lead").get("he"):
             d["lead"] = kwargs.get("lead")
         return d
 
@@ -871,7 +871,6 @@ class GroupSheetListFactory(AbstractStoryFactory):
         create_israel_and_diaspora_stories(_create_nechama_sheet_story, **kwargs)
 
 
-
 class SheetListFactory(AbstractStoryFactory):
     """
         "title" : {
@@ -907,7 +906,7 @@ class SheetListFactory(AbstractStoryFactory):
             "title": title,
         }
 
-        if kwargs.get("lead"):
+        if kwargs.get("lead") and kwargs.get("lead").get("en") and kwargs.get("lead").get("he"):
             d["lead"] = kwargs.get("lead")
 
         return d
@@ -918,7 +917,7 @@ class SheetListFactory(AbstractStoryFactory):
 
     @classmethod
     def _get_featured_ids(cls, k):
-        shts = db.sheets.find({"is_featured": True}, {"id":1})
+        shts = db.sheets.find({"is_featured": True}, {"id": 1})
         ids = [s["id"] for s in shts]
         return random.sample(ids, k)
 
@@ -1012,9 +1011,9 @@ class TopicListStoryFactory(AbstractStoryFactory):
 
     @classmethod
     def create_trending_story(cls, **kwargs):
-        days = kwargs.get("days", 14)
+        days = kwargs.get("days", 7)
         from sefaria import sheets
-        topics = [t["tag"] for t in sheets.recent_public_tags(days=days, ntags=6)]
+        topics = [t["tag"] for t in sheets.trending_tags(days=days, ntags=6)]
         cls.create_shared_story(topics=topics)
 
     @classmethod
@@ -1098,7 +1097,7 @@ class TopicTextsStoryFactory(AbstractStoryFactory):
     def generate_random_shared_story(cls, **kwargs):
         from . import topic
 
-        topics_filtered = filter(lambda x: x['count'] > 15, topic.get_topics().list())
+        topics_filtered = filter(lambda x: x['good_to_promote'], topic.get_topics().list())
         random_topic = random.choice(topics_filtered)['tag']
 
         return cls._generate_shared_story(topic=random_topic, **kwargs)
