@@ -471,7 +471,7 @@ class TextIndexer(object):
                 raise e
 
     @classmethod
-    def index_all(cls, index_name, merged=False, debug=False):
+    def index_all(cls, index_name, merged=False, debug=False, for_es=True, action=None):
         cls.index_name = index_name
         cls.merged = merged
         cls.create_version_priority_map()
@@ -493,26 +493,30 @@ class TextIndexer(object):
         versions = None  # release RAM
         for title, vlist in versions_by_index.items():
             cls.trefs_seen = set()
-            cls._bulk_actions = []
             cls.curr_index = vlist[0].get_index() if len(vlist) > 0 else None
-            try:
-                cls.best_time_period = cls.curr_index.best_time_period()
-            except ValueError:
-                cls.best_time_period = None
+            if for_es:
+                cls._bulk_actions = []
+                try:
+                    cls.best_time_period = cls.curr_index.best_time_period()
+                except ValueError:
+                    cls.best_time_period = None
             for v in vlist:
                 if v.versionTitle == u"Yehoyesh's Yiddish Tanakh Translation [yi]":
                     print "skipping yiddish. we don't like yiddish"
                     continue
 
-                cls.index_version(v)
+                cls.index_version(v, action=action)
                 print "Indexed Version {}/{}".format(vcount, total_versions)
                 vcount += 1
-            bulk(es_client, cls._bulk_actions, stats_only=True, raise_on_error=False)
+            if for_es:
+                bulk(es_client, cls._bulk_actions, stats_only=True, raise_on_error=False)
 
     @classmethod
-    def index_version(cls, version, tries=0):
+    def index_version(cls, version, tries=0, action=None):
+        if not action:
+            action = cls._cache_action
         try:
-            version.walk_thru_contents(cls._cache_action, heTref=cls.curr_index.get_title('he'), schema=cls.curr_index.schema, terms_dict=cls.terms_dict)
+            version.walk_thru_contents(action, heTref=cls.curr_index.get_title('he'), schema=cls.curr_index.schema, terms_dict=cls.terms_dict)
         except pymongo.errors.AutoReconnect as e:
             # Adding this because there is a mongo call for dictionary words in walk_thru_contents()
             if tries < 200:
