@@ -42,56 +42,61 @@ def refresh_test():
 
 
 def ensure_indices():
-    db.following.ensure_index("follower")
-    db.following.ensure_index("followee")
-    db.history.ensure_index("revision")
-    db.history.ensure_index("method")
-    db.history.ensure_index([("ref", pymongo.ASCENDING), ("version", pymongo.ASCENDING), ("language", pymongo.ASCENDING)])
-    db.history.ensure_index("date")
-    db.history.ensure_index("ref")
-    db.history.ensure_index("user")
-    db.history.ensure_index("rev_type")
-    db.history.ensure_index("version")
-    db.history.ensure_index("new.refs")
-    db.history.ensure_index("new.ref")
-    db.history.ensure_index("old.refs")
-    db.history.ensure_index("old.ref")
-    db.history.ensure_index("title")
-    db.index.ensure_index("title")
-    db.index_queue.ensure_index([("lang", pymongo.ASCENDING), ("version", pymongo.ASCENDING), ("ref", pymongo.ASCENDING)], unique=True)
-    db.links.ensure_index("refs")
-    db.links.ensure_index("refs.0")
-    db.links.ensure_index("refs.1")
-    db.links.ensure_index("source_text_oid")
-    db.links.ensure_index("is_first_comment")
-    db.metrics.ensure_index("timestamp", unique=True)
-    db.notes.ensure_index([("owner", pymongo.ASCENDING), ("ref", pymongo.ASCENDING), ("public", pymongo.ASCENDING)])
-    db.notifications.ensure_index([("uid", pymongo.ASCENDING), ("read", pymongo.ASCENDING)])
-    db.notifications.ensure_index("uid")
-    db.parshiot.ensure_index("date")
-    db.place.ensure_index([("point", pymongo.GEOSPHERE)])
-    db.place.ensure_index([("area", pymongo.GEOSPHERE)])
-    db.person.ensure_index("key")
-    db.profiles.ensure_index("slug")
-    db.profiles.ensure_index("id")
-    db.sheets.ensure_index("id")
-    db.sheets.ensure_index("dateModified")
-    db.sheets.ensure_index("sources.ref")
-    db.sheets.ensure_index("includedRefs")
-    db.sheets.ensure_index("tags")
-    db.sheets.ensure_index("owner")
-    db.sheets.ensure_index("assignment_id")
-    db.sheets.ensure_index("is_featured")
-    db.texts.ensure_index("title")
-    db.texts.ensure_index([("priority", pymongo.DESCENDING), ("_id", pymongo.ASCENDING)])
-    db.texts.ensure_index([("versionTitle", pymongo.ASCENDING), ("langauge", pymongo.ASCENDING)])
-    db.word_form.ensure_index("form")
-    db.word_form.ensure_index("c_form")
-    db.term.ensure_index("titles.text", unique=True)
-    db.lexicon_entry.ensure_index([("headword", pymongo.ASCENDING), ("parent_lexicon", pymongo.ASCENDING)])
-    db.user_story.ensure_index([("uid", pymongo.ASCENDING), ("timestamp", pymongo.DESCENDING)])
-    db.user_story.ensure_index([("timestamp", pymongo.DESCENDING)])
-    db.passage.ensure_index("ref_list")
-    db.user_history.ensure_index("uid")
-    db.trend.ensure_index("name")
-    db.trend.ensure_index("uid")
+    db.following.create_index("follower")
+    db.following.create_index("followee")
+    db.history.create_index("revision")
+    db.history.create_index("method")
+    db.history.create_index([("ref", pymongo.ASCENDING), ("version", pymongo.ASCENDING), ("language", pymongo.ASCENDING)])
+    db.history.create_index("date")
+    db.history.create_index("ref")
+    db.history.create_index("user")
+    db.history.create_index("rev_type")
+    db.history.create_index("version")
+    db.history.create_index("new.refs")
+    db.history.create_index("new.ref")
+    db.history.create_index("old.refs")
+    db.history.create_index("old.ref")
+    db.history.create_index("title")
+    db.index.create_index("title")
+    db.index_queue.create_index([("lang", pymongo.ASCENDING), ("version", pymongo.ASCENDING), ("ref", pymongo.ASCENDING)], unique=True)
+    db.links.create_index("refs")
+    db.links.create_index("refs.0")
+    db.links.create_index("refs.1")
+    db.links.create_index("expandedRefs0")
+    db.links.create_index("expandedRefs1")
+    db.links.create_index("source_text_oid")
+    db.links.create_index("is_first_comment")
+    db.metrics.create_index("timestamp", unique=True)
+    db.notes.create_index([("owner", pymongo.ASCENDING), ("ref", pymongo.ASCENDING), ("public", pymongo.ASCENDING)])
+    db.notifications.create_index([("uid", pymongo.ASCENDING), ("read", pymongo.ASCENDING)])
+    db.notifications.create_index("uid")
+    db.parshiot.create_index("date")
+    db.place.create_index([("point", pymongo.GEOSPHERE)])
+    db.place.create_index([("area", pymongo.GEOSPHERE)])
+    db.person.create_index("key")
+    db.profiles.create_index("slug")
+    db.profiles.create_index("id")
+    db.sheets.create_index("id")
+    db.sheets.create_index("dateModified")
+    db.sheets.create_index("sources.ref")
+    db.sheets.create_index("includedRefs")
+    db.sheets.create_index("tags")
+    db.sheets.create_index("owner")
+    db.sheets.create_index("assignment_id")
+    db.sheets.create_index("is_featured")
+    db.texts.create_index("title")
+    db.texts.create_index([("priority", pymongo.DESCENDING), ("_id", pymongo.ASCENDING)])
+    db.texts.create_index([("versionTitle", pymongo.ASCENDING), ("langauge", pymongo.ASCENDING)])
+    db.word_form.create_index("form")
+    db.word_form.create_index("c_form")
+    db.term.create_index("titles.text", unique=True)
+    db.lexicon_entry.create_index([("headword", pymongo.ASCENDING), ("parent_lexicon", pymongo.ASCENDING)])
+    db.user_story.create_index([("uid", pymongo.ASCENDING), ("timestamp", pymongo.DESCENDING)])
+    db.user_story.create_index([("timestamp", pymongo.DESCENDING)])
+    db.passage.create_index("ref_list")
+    db.user_history.create_index("uid")
+    db.user_history.create_index("sheet_id")
+    db.user_history.create_index("datetime")
+    db.trend.create_index("name")
+    db.trend.create_index("uid")
+    db.webpages.create_index("ref")
