@@ -11,9 +11,16 @@ def setup_module(module):
 class Test_get_links():
 
     def test_get_links_on_range(self):
-        x = len(get_links("Exodus 2:3"))
-        y = len(get_links("Exodus 2:4"))
-        assert len(get_links("Exodus 2:3-4")) == (x+y)
+        r3 = [l["ref"] + l["type"] for l in get_links("Exodus 2:3")]
+        r4 = [l["ref"] + l["type"]  for l in get_links("Exodus 2:4")]
+        r34 = [l["ref"] + l["type"]  for l in get_links("Exodus 2:3-4")]
+
+        # All links in first segment present in range
+        assert all([r in r34 for r in r3])
+        # All links in second segment present in range
+        assert all([r in r34 for r in r4])
+        # No links in range absent from segments
+        assert all(r in r3 or r in r4 for r in r34)
 
 
 class Test_links_from_get_text():
