@@ -22,9 +22,11 @@ const AccountPanel              = require('./AccountPanel');
 const NotificationsPanel        = require('./NotificationsPanel');
 const MyNotesPanel              = require('./MyNotesPanel');
 const UserHistoryPanel          = require('./UserHistoryPanel');
+const UserProfile               = require('./UserProfile');
 const UpdatesPanel              = require('./UpdatesPanel');
 const HomeFeed                  = require('./HomeFeed');
 const StoryEditor               = require('./StoryEditor');
+const UserStats                 = require('./UserStats');
 const ModeratorToolsPanel       = require('./ModeratorToolsPanel');
 const {
   MyGroupsPanel,
@@ -103,6 +105,7 @@ class ReaderPanel extends Component {
       displaySettingsOpen:  false,
       tagSort: "count",
       mySheetSort: "date",
+      profile: props.initialProfile || null,
       initialAnalyticsTracked: false
     }
   }
@@ -624,6 +627,7 @@ class ReaderPanel extends Component {
           contentLang={this.state.settings.language}
           interfaceLang={this.props.interfaceLang}
           onSegmentClick={this.handleSheetSegmentClick}
+          openProfile={this.props.openProfile}
       />);
     }
     if (this.state.mode === "Text" || this.state.mode === "TextAndConnections") {
@@ -716,6 +720,7 @@ class ReaderPanel extends Component {
           setVersionFilter={this.setVersionFilter}
           viewExtendedNotes={this.props.viewExtendedNotes.bind(null, "Connections")}
           checkIntentTimer={this.props.checkIntentTimer}
+          openProfile={this.props.openProfile}
           key="connections" />
       );
     }
@@ -844,6 +849,7 @@ class ReaderPanel extends Component {
                     updateAppliedOptionField={this.props.updateSearchOptionField}
                     updateAppliedOptionSort={this.props.updateSearchOptionSort}
                     registerAvailableFilters={this.props.registerAvailableFilters}
+                    openProfile={this.props.openProfile}
                   />);
 
     } else if (this.state.menuOpen === "sheets") {
@@ -864,7 +870,10 @@ class ReaderPanel extends Component {
                     setSheetTag={this.setSheetTag}
                     setGroupTag={this.setGroupTag}
                     searchInGroup={this.props.searchInGroup}
-                    key={"SheetsNav"} />);
+                    openProfile={this.props.openProfile}
+                    key={"SheetsNav"}
+                    openProfile={this.props.openProfile}
+                  />);
 
     } else if (this.state.menuOpen === "topics") {
       if (this.state.navigationTopic) {
@@ -948,8 +957,11 @@ class ReaderPanel extends Component {
 
     } else if (this.state.menuOpen === "story_editor") {
       menu = (<StoryEditor
+                    toggleSignUpModal={this.props.toggleSignUpModal}
                     interfaceLang={this.props.interfaceLang}
       />);
+    } else if (this.state.menuOpen === "user_stats") {
+      menu = (<UserStats/>);
 
     /* todo: do we need this here?
                     multiPanel={this.props.multiPanel}
@@ -972,7 +984,16 @@ class ReaderPanel extends Component {
           hideNavHeader={this.props.hideNavHeader}
           interfaceLang={this.props.interfaceLang}
         />
-      )
+      );
+    } else if (this.state.menuOpen === "profile") {
+      menu = (
+        <UserProfile
+          profile={this.state.profile}
+          handleInAppLinkClick={this.props.handleInAppLinkClick}
+          openProfile={this.props.openProfile}
+          toggleSignUpModal={this.props.toggleSignUpModal}
+        />
+      );
     }
 
     let classes  = {readerPanel: 1, narrowColumn: this.state.width < 730};
@@ -1065,6 +1086,7 @@ ReaderPanel.propTypes = {
   initialSheetSearchField:          PropTypes.string,
   initialSheetSearchSortType:       PropTypes.string,
   initialSheetsTag:            PropTypes.string,
+  initialProfile:              PropTypes.object,
   initialState:                PropTypes.object, // if present, overrides all props above
   interfaceLang:               PropTypes.string,
   setCentralState:             PropTypes.func,
@@ -1111,6 +1133,8 @@ ReaderPanel.propTypes = {
   checkIntentTimer:            PropTypes.func,
   toggleSignUpModal:           PropTypes.func.isRequired,
   getHistoryRef:               PropTypes.func,
+  profile:                     PropTypes.object,
+  openProfile:                 PropTypes.func,
 };
 
 
