@@ -304,6 +304,7 @@ class SearchResultList extends Component {
               <SearchSheetResult
                     data={result}
                     query={this.props.query}
+                    openProfile={this.props.openProfile}
                     key={result._id} />);
         }
 
@@ -354,7 +355,8 @@ SearchResultList.propTypes = {
   updateAppliedFilter:      PropTypes.func,
   updateAppliedOptionField: PropTypes.func,
   updateAppliedOptionSort:  PropTypes.func,
-  registerAvailableFilters: PropTypes.func
+  registerAvailableFilters: PropTypes.func,
+  openProfile:              PropTypes.func.isRequired,
 };
 
 
