@@ -1,6 +1,7 @@
 const {
   LoadingMessage,
   ReaderMessage,
+  ProfilePic,
 } = require('./Misc');
 
 const React = require('react');
@@ -93,6 +94,7 @@ class Sheet extends Component {
             hasSidebar = {this.props.hasSidebar}
             sheetNumbered = {sheet.options.numbered}
             sheetID = {sheet.id}
+            openProfile={this.props.openProfile}
           />
       )
     }
@@ -198,6 +200,13 @@ class SheetContent extends Component {
     this.props.onRefClick(ref);
   }
 
+  openProfile(e) {
+    e.preventDefault();
+    const slugMatch = this.props.authorUrl.match(/profile\/(.+)$/);
+    const slug = !!slugMatch ? slugMatch[1] : '';
+    this.props.openProfile(slug, this.props.authorStatement);
+  }
+
   render() {
     var sources = this.props.sources.length ? this.props.sources.map(function(source, i) {
       const highlightedRef = this.props.highlightedRefsInSheet ? Sefaria.normRefList(this.props.highlightedRefsInSheet) : null;
@@ -294,11 +303,15 @@ class SheetContent extends Component {
 
             <div className="authorStatement">
                 <div className="groupListingImageBox imageBox">
-                    <a href={this.props.authorUrl}>
-                        <img className="groupListingImage img-circle" src={this.props.authorImage} alt="Author Avatar" />
+                    <a href={this.props.authorUrl} onClick={this.openProfile}>
+                      <ProfilePic
+                        url={this.props.authorImage}
+                        len={30}
+                        name={this.props.authorStatement}
+                      />
                     </a>
                 </div>
-                <span>by <a href={this.props.authorUrl}>{this.props.authorStatement}</a></span>
+                <span>by <a href={this.props.authorUrl} onClick={this.openProfile}>{this.props.authorStatement}</a></span>
             </div>
 
             {this.props.group && this.props.group != "" ?
