@@ -95,7 +95,6 @@ class TextList extends Component {
       this.setState({waitForText: false});
       return;
     }
-
     this.setState({waitForText: true});
     Sefaria.text(commentarySection, {}, function() {
       if (this._isMounted) {
@@ -175,7 +174,10 @@ class TextList extends Component {
     }.bind(this);
 
     let sectionLinks = Sefaria.getLinksFromCache(sectionRef);
-    let overlaps = link => (!(Sefaria.splitRangingRef(link.anchorRef).every(aref => Sefaria.util.inArray(aref, refs) === -1)));
+    sectionLinks.map(link => {
+      if (!("anchorRefExpanded" in link)) { link.anchorRefExpanded = Sefaria.splitRangingRef(link.anchorRef); }
+    });
+    let overlaps = link => (!(link.anchorRefExpanded.every(aref => Sefaria.util.inArray(aref, refs) === -1)));
     let links = Sefaria._filterLinks(sectionLinks, filter)
       .filter(overlaps)
       .sort(sortConnections);
